Introduce BillingDetailsEntry
